change Volume to be an int instead of a float

+++ src/backtest_lib/universe/__init__.py
@@ -16,7 +16,7 @@ R = TypeVar("R", bound=Real, contravariant=True)
 
 SecurityName = str
 Price = float
-Volume = float
+Volume = int
 
 
 PeriodIndex = Comparable
